COR-223: The list of users does not appear in the Community Management portlet when using Oracle Virtual Directory LDAP
What is the problem to fix?
When we use OpenLDAP 2.4-20, the list of users appears in the portlet "Community Management". 
However, when we connect Oracle Virtual Directory to OpenLDAP 2.4-20 through Oracle Virtual Directory Manager, this list no longer appears.
When we debug the class "LDAPUserPageList" that is available in the package "org.exoplatform.services.organization.ldap", we note that "responseControls" has the value "null", which was not predicted in the code. 

Fix description
* Skip when responseControls[] array is null.